price type changed to Float

+++ src/tokens/dto/buy-token.input.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int } from '@nestjs/graphql';
+import { Field, Float, InputType, Int } from '@nestjs/graphql';
 import { IsNotEmpty, IsNumber } from 'class-validator';
 
 @InputType()
@@ -8,7 +8,7 @@ export class BuyTokenInput {
   @IsNotEmpty()
   id: number;
 
-  @Field(() => Int)
+  @Field(() => Float)
   @IsNumber()
   @IsNotEmpty()
   price: number;
